fixed readme and adc.h function declaration

+++ adc.h
@@ -1,2 +1,2 @@
 void adc_init(void);
-uint8_t adc_sample(uint8_t);
+uint16_t adc_sample(uint8_t);
